Support decimal series volume numbers (e.g. 4.5, 9.5)
Book.seriesVolume is now a double so half-volumes are supported. The
edit screen's volume field accepts a decimal point, and both the edit
and detail screens format whole numbers without a trailing ".0" via a
new Book.seriesVolumeDisplay getter.

+++ lib/services/database_service.dart
@@ -38,7 +38,7 @@ class DatabaseService {
             authors TEXT,
             illustrators TEXT,
             series TEXT,
-            seriesVolume INTEGER,
+            seriesVolume REAL,
             publisher TEXT,
             publishedDate TEXT,
             description TEXT,
